New sections about UIO and DUIF3.
We have added two new sections about _User Interface Organizer_ (UIO), and
_DarNified UI F3_ (DUIF3).

UIO is a FOSE plugin that serves as a structure keeper for every UI mod we
install.

DUIF3 is a revamp of the user interface to make it more palatable for PC
users.

Additionaly, we have added a few screenshots to further extend the
information about CASM.