I have corrected the inner structure of some of the texture mods in this chapter, as I did with chapter 6.4.
Taking a look at the data folder of _Fallout 3_, I can see several folders and files that hardly fit in there. I will have to scan through all the installed mods, looking for strange structures to fix them, if necessary.

For now, I am going to fix chapter 6.9, because the texture mod for _Mothership Zeta_ is neither suitable for automatic installation with NMM.