The mod _Mothership Zeta DLC Retextured_ had a NMM non-compatible structure, so I have had to fix it and reinstall it.
And that's it. This is the last piece of the puzzle. I consider the contents of the guide as closed.

But there are still a lot of things to do, beginning with a restructure of the site, maybe with a wiki and a download link (at least for the HTML version, for the time being).

Stay tuned!