I have automated the compilation process of the guide (sort of, at least).
Using Maven and the Asciidoc maven plugin, I have created a POM file to create the final HTML5 product. There are a few quirks, anyway, because I am using BashOnWindows to install the toolchain, and there are thing a little bit underdeveloped right now.

The structure of the project has changed dramatically, in order to give it a more software-like approach. All the contents are now under `[src/main]` folder, and the final product goes to the `[build]` folder.