Final version of the guide.
Yeah, this is the final version of the guide. At last.

I have added a new chapter (really), with a reference to a saved game I created using just the mods from chapter 4 (_Fixing the game_). I saved the game right after leaving Vault 101, so that you can customize your character the way you fuckin' want.

I also added a file I completely missed in the installation of FWE, regarding the unofficial patch.

There will be just another version of the guide with an improved building system using Maven Assembly Plugin, but this is the final guide, content-wise.

Now, let's focus on the second volume `:P`